requirements.txt
solana
solders
aiohttp
python-dotenv
base58